requirements.txt
mcp-python-sdk>=1.0.0
httpx>=0.24.0
asyncio>=3.4.3